Demo multi context in single-threaded app.
